+++ src/pdfCropMargins/external_program_calls.py
@@ -210,8 +210,8 @@ def remove_program_temp_directory():
 def cleanup_and_exit(exit_code, stack_frame=None):
     """Exit the program, after cleaning up the temporary directory.  The `stack_frame`
     argument is for when `signal.signal` calls the function."""
-    #if stack_frame is not None:
-    #    print("Signal caught.")
+    if stack_frame is not None:
+        print("\nProcess killed by a signal...", file=sys.stderr)
     remove_program_temp_directory()
     sys.exit(exit_code)